requirements.txt
asgiref==3.4.1
certifi==2022.9.24
charset-normalizer==2.0.12
colorama==0.4.5
defusedxml==0.7.1
Django==3.2.16
djangorestframework==3.14.0
httpie==2.6.0
idna==3.4
Pygments==2.13.0
PySocks==1.7.1
pytz==2022.6
requests==2.27.1
requests-toolbelt==0.10.1
sqlparse==0.4.3
typing_extensions==4.1.1
urllib3==1.26.13
